// generated/src/app/api-models/pmb-snap-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbSnapWork {
  seqGpbilId: number;
  seqSubsId: number;
  seqSubsUpdate: number;
  billDetailFromDate: Date;
  billDetailThruDate: Date;
  seqGroupId: number;
  planCode: string;
  familySize: number;
  spouseFlag: string;
  medicareStatus: string;
  otherStatusFlag: string;
  subscriberDob: Date;
  subscriberGender: string;
  salary: number;
  premOverrideStep: number;
  premOverrideAmt: number;
  premOverrideCode: string;
  customerType: string;
  customerId: string;
  positivePremium: string;
  prorationMethod: string;
  billEffectiveFrom: Date;
  subscDept: string;
  subscLocation: string;
  seqParentId: number;
  rateType: string;
}